+++ TruncatedMeanCopy/Algorithm/TruncMean.h
@@ -32,7 +32,7 @@
    doxygen documentation!
 */
 
-static const float kINVALID_FLOAT = std::numeric_limits<float>::max();
+static const float kINVALID_FLOAT = std::numeric_limits<double>::max();
 
 class TruncMean{
 
@@ -55,18 +55,18 @@ class TruncMean{
      1) the median and rms of these values is calculated.
      2) the subset of local dq values within the range [median-rms, median+rms] is selected.
      3) the resulting local truncated dq is the average of this truncated subset.
-     @input std::vector<float> rr_v -> vector of x-axis coordinates (i.e. position for track profile)
-     @input std::vector<float> dq_v -> vector of measured values for which truncated profile is requested
+     @input std::vector<double> rr_v -> vector of x-axis coordinates (i.e. position for track profile)
+     @input std::vector<double> dq_v -> vector of measured values for which truncated profile is requested
      (i.e. charge profile of a track)
-     @input std::vector<float> dq_trunc_v -> passed by reference -> output stored here
+     @input std::vector<double> dq_trunc_v -> passed by reference -> output stored here
      @input float nsigma -> optional parameter, number of sigma to keep around RMS for TM calculation
   */
-  void CalcTruncMeanProfile(const std::vector<float>& rr_v, const std::vector<float>& dq_v,
-			    std::vector<float>& dq_trunc_v, const float& nsigma = 1);
+  void CalcTruncMeanProfile(const std::vector<double>& rr_v, const std::vector<double>& dq_v,
+			    std::vector<double>& dq_trunc_v, const float& nsigma = 1);
 
   /**
      @brief Iteratively calculate the truncated mean of a distribution
-     @input std::vector<float> v -> vector of values for which truncated mean is asked
+     @input std::vector<double> v -> vector of values for which truncated mean is asked
      @input size_t nmin -> minimum number of iterations to converge on truncated mean
      @input size_t nmax -> maximum number of iterations to converge on truncated mean
      @input size_t lmin -> minimum number of entries in vector before exiting and returning current value
@@ -88,9 +88,9 @@ class TruncMean{
 
  private:
 
-  float Mean  (const std::vector<float>& v);
-  float Median(const std::vector<float>& v);
-  float RMS   (const std::vector<float>& v);
+  float Mean  (const std::vector<double>& v);
+  float Median(const std::vector<double>& v);
+  float RMS   (const std::vector<double>& v);
 
   /**
      Smearing radius over which charge from neighboring hits is scanned to calculate local
